Year 2/Semester 2/Systems for Design and Implementation/LAB5/backend_django/employee/Views/TaskViews.py
from rest_framework import generics
from employee.models import Task
from employee.serializers import TaskSerializer2




class TaskList(generics.ListCreateAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer2


class TaskDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Task.objects.all()
    serializer_class = TaskSerializer2
